Make error message creation on login reusable
