+++ packages/errors/src/index.ts
@@ -2,5 +2,6 @@ export { BaseError } from './BaseError';
 export { Problem } from './Problem';
 export { ERROR_TYPE, type ErrorType } from './constants';
 export { isRecoverable, extractErrorFromFetchResponse } from './extractErrorFromFetchResponse';
+export { getLoginErrorMessage } from './getLoginErrorMessage';
 
 export * from './errors';
